Factor out classes from `Button` component
